--- backend/src/main/java/moheng/planner/Plan.java
@@ -1,36 +1,0 @@
-package moheng.planner;
-
-import jakarta.persistence.*;
-import moheng.member.domain.Member;
-
-import java.time.LocalDate;
-
-@Table(name = "plan")
-@Entity
-public class Plan {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "name", nullable = false)
-    private String name;
-
-    @Column(name = "start_date", nullable = false)
-    private LocalDate startDate;
-
-    @Column(name = "end_date", nullable = false)
-    private LocalDate endDate;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    protected Plan() {
-    }
-
-    public Plan(final String name, final LocalDate startDate, final LocalDate endDate) {
-        this.name = name;
-        this.startDate = startDate;
-        this.endDate = endDate;
-    }
-}
