+++ backend/src/main/java/moheng/planner/domain/Plan.java
@@ -1,0 +1,55 @@
+package moheng.planner.domain;
+
+import jakarta.persistence.*;
+import moheng.member.domain.Member;
+import moheng.planner.exception.InvalidPlanDateException;
+import moheng.planner.exception.InvalidPlanNameException;
+
+import java.time.LocalDate;
+
+@Table(name = "plan")
+@Entity
+public class Plan {
+    private static final int MIN_NAME_LENGTH = 2;
+    private static final int MAX_NAME_LENGTH = 100;
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "start_date", nullable = false)
+    private LocalDate startDate;
+
+    @Column(name = "end_date", nullable = false)
+    private LocalDate endDate;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    protected Plan() {
+    }
+
+    public Plan(final String name, final LocalDate startDate, final LocalDate endDate, final Member member) {
+        validateName(name);
+        validateDate(startDate, endDate);
+        this.name = name;
+        this.startDate = startDate;
+        this.endDate = endDate;
+    }
+
+    private void validateName(final String name) {
+        if(name.length() < MIN_NAME_LENGTH || name.length() > MAX_NAME_LENGTH) {
+            throw new InvalidPlanNameException(String.format("이름은 %d자 이상 1자 %d이하여야 합니다.",  MIN_NAME_LENGTH, MAX_NAME_LENGTH));
+        }
+    }
+
+    private void validateDate(final LocalDate startDate, final LocalDate endDate) {
+        if(startDate.isAfter(endDate)) {
+            throw new InvalidPlanDateException("플랜의 시작날짜가 종료날짜보다 늦을 수 없습니다.");
+        }
+    }
+}
