+++ backend/src/main/java/moheng/planner/exception/InvalidPlanDateException.java
@@ -1,0 +1,7 @@
+package moheng.planner.exception;
+
+public class InvalidPlanDateException extends RuntimeException {
+    public InvalidPlanDateException(final String message) {
+        super(message);
+    }
+}
