+++ backend/src/main/java/moheng/planner/exception/InvalidPlanNameException.java
@@ -1,0 +1,7 @@
+package moheng.planner.exception;
+
+public class InvalidPlanNameException extends RuntimeException {
+    public InvalidPlanNameException(final String message) {
+        super(message);
+    }
+}
